requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.12
